This stuff is rotting and I gotta rescue it, popups or not.

Also I could be wrong (I hope not) but on Weston for now the editor's
working fine (AFTER verifying that I had disabled single-window mode by
making the window as small as possible and opening a popup menu), so
perhaps that "earlier-than-sway setting" bug was just an edge case or
something minor than I thought. Still, I'll wait a bit before calling
this a sway only problem because it's way past midnight and I'm tired,
making me extremely vulnerable to dumbass mistakes that illude me that
I can finally finish this port without stumbling on yet other
compositor problems (as we all know by know).

+++ tests/core/math/test_astar.h
@@ -58,7 +58,7 @@ public:
 	}
 
 	// Disable heuristic completely.
-	real_t _compute_cost(int p_from, int p_to) {
+	real_t _compute_cost(int64_t p_from, int64_t p_to) {
 		if (p_from == A && p_to == C) {
 			return 1000;
 		}
@@ -68,7 +68,7 @@ public:
 
 TEST_CASE("[AStar3D] ABC path") {
 	ABCX abcx;
-	Vector<int> path = abcx.get_id_path(ABCX::A, ABCX::C);
+	Vector<int64_t> path = abcx.get_id_path(ABCX::A, ABCX::C);
 	REQUIRE(path.size() == 3);
 	CHECK(path[0] == ABCX::A);
 	CHECK(path[1] == ABCX::B);
@@ -77,7 +77,7 @@ TEST_CASE("[AStar3D] ABC path") {
 
 TEST_CASE("[AStar3D] ABCX path") {
 	ABCX abcx;
-	Vector<int> path = abcx.get_id_path(ABCX::X, ABCX::C);
+	Vector<int64_t> path = abcx.get_id_path(ABCX::X, ABCX::C);
 	REQUIRE(path.size() == 4);
 	CHECK(path[0] == ABCX::X);
 	CHECK(path[1] == ABCX::A);
@@ -318,7 +318,7 @@ TEST_CASE("[Stress][AStar3D] Find paths") {
 		for (int u = 0; u < N; u++) {
 			for (int v = 0; v < N; v++) {
 				if (u != v) {
-					Vector<int> route = a.get_id_path(u, v);
+					Vector<int64_t> route = a.get_id_path(u, v);
 					if (!Math::is_inf(d[u][v])) {
 						// Reachable.
 						if (route.size() == 0) {
